Extract schedule helper for receivers
